backend/pyproject.toml
[project]
name = "Filme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.24.0",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
]
